Fixed a bug with time in milliseconds being cast to int (thanks to Mihail)

See merge request cse1105/2019-2020/repos/oopp-group-28/repository-template!98

+++ server/src/main/java/nl/tudelft/oopp/demo/services/RoomReservationService.java
@@ -72,10 +72,10 @@ public class RoomReservationService {
 
         switch (attribute) {
             case "fromDate":
-                roomReservation.setFromTime(new Date(Integer.parseInt(value)));
+                roomReservation.setFromTime(new Date(Long.parseLong(value)));
                 break;
             case "toDate":
-                roomReservation.setToTime(new Date(Integer.parseInt(value)));
+                roomReservation.setToTime(new Date(Long.parseLong(value)));
                 break;
             case "roomId":
                 int roomId = Integer.parseInt(value);
